+++ crates/wdl-format/src/v1.rs
@@ -338,11 +338,17 @@ pub fn format_literal_input(
         }
     }
 
+    let mut items = items.iter().peekable();
     let mut commas = commas.iter();
-    for item in items {
-        (&item).write(stream, config);
+    while let Some(item) = items.next() {
+        (item).write(stream, config);
         if let Some(comma) = commas.next() {
-            (comma).write(stream, config);
+            // check if this comma can be dropped when trailing commas are disabled. Comma
+            // can be dropped iff this is the last item and the comma does not
+            // have a comment.
+            if config.trailing_commas || items.peek().is_some() || comma.has_comment() {
+                (comma).write(stream, config);
+            }
         } else if config.trailing_commas {
             stream.push_literal(",".to_string(), SyntaxKind::Comma);
         }
@@ -415,11 +421,17 @@ pub fn format_literal_hints(
         }
     }
 
+    let mut items = items.iter().peekable();
     let mut commas = commas.iter();
-    for item in items {
-        (&item).write(stream, config);
+    while let Some(item) = items.next() {
+        (item).write(stream, config);
         if let Some(comma) = commas.next() {
-            (comma).write(stream, config);
+            // check if this comma can be dropped when trailing commas are disabled. Comma
+            // can be dropped iff this is the last item and the comma does not
+            // have a comment.
+            if config.trailing_commas || items.peek().is_some() || comma.has_comment() {
+                (comma).write(stream, config);
+            }
         } else if config.trailing_commas {
             stream.push_literal(",".to_string(), SyntaxKind::Comma);
         }
@@ -494,11 +506,17 @@ pub fn format_literal_output(
         }
     }
 
+    let mut items = items.iter().peekable();
     let mut commas = commas.iter();
-    for item in items {
-        (&item).write(stream, config);
+    while let Some(item) = items.next() {
+        (item).write(stream, config);
         if let Some(comma) = commas.next() {
-            (comma).write(stream, config);
+            // check if this comma can be dropped when trailing commas are disabled. Comma
+            // can be dropped iff this is the last item and the comma does not
+            // have a comment.
+            if config.trailing_commas || items.peek().is_some() || comma.has_comment() {
+                (comma).write(stream, config);
+            }
         } else if config.trailing_commas {
             stream.push_literal(",".to_string(), SyntaxKind::Comma);
         }