+++ crates/wdl-format/src/v1/expr.rs
@@ -392,11 +392,18 @@ pub fn format_literal_array(
     if !empty {
         stream.increment_indent();
     }
+
+    let mut items = items.iter().peekable();
     let mut commas = commas.iter();
-    for item in items {
-        (&item).write(stream, config);
+    while let Some(item) = items.next() {
+        (item).write(stream, config);
         if let Some(comma) = commas.next() {
-            (comma).write(stream, config);
+            // check if this comma can be dropped when trailing commas are disabled. Comma
+            // can be dropped iff this is the last item and the comma does not
+            // have a comment.
+            if config.trailing_commas || items.peek().is_some() || comma.has_comment() {
+                (comma).write(stream, config);
+            }
         } else if config.trailing_commas {
             stream.push_literal(",".to_string(), SyntaxKind::Comma);
         }
@@ -468,11 +475,17 @@ pub fn format_literal_map(
         }
     }
 
+    let mut items = items.iter().peekable();
     let mut commas = commas.iter();
-    for item in items {
-        (&item).write(stream, config);
+    while let Some(item) = items.next() {
+        (item).write(stream, config);
         if let Some(comma) = commas.next() {
-            (comma).write(stream, config);
+            // check if this comma can be dropped when trailing commas are disabled. Comma
+            // can be dropped iff this is the last item and the comma does not
+            // have a comment.
+            if config.trailing_commas || items.peek().is_some() || comma.has_comment() {
+                (comma).write(stream, config);
+            }
         } else if config.trailing_commas {
             stream.push_literal(",".to_string(), SyntaxKind::Comma);
         }
@@ -549,11 +562,17 @@ pub fn format_literal_object(
         }
     }
 
+    let mut items = members.iter().peekable();
     let mut commas = commas.iter();
-    for member in members {
-        (&member).write(stream, config);
+    while let Some(item) = items.next() {
+        (item).write(stream, config);
         if let Some(comma) = commas.next() {
-            (comma).write(stream, config);
+            // check if this comma can be dropped when trailing commas are disabled. Comma
+            // can be dropped iff this is the last item and the comma does not
+            // have a comment.
+            if config.trailing_commas || items.peek().is_some() || comma.has_comment() {
+                (comma).write(stream, config);
+            }
         } else if config.trailing_commas {
             stream.push_literal(",".to_string(), SyntaxKind::Comma);
         }