+++ crates/wdl-format/src/v1/workflow.rs
@@ -303,17 +303,19 @@ pub fn format_workflow_hints_array(
         }
     }
 
-    let mut commas = commas.into_iter();
-    for item in items {
-        (&item).write(stream, config);
-        match commas.next() {
-            Some(comma) => {
-                (&comma).write(stream, config);
+    let mut items = items.iter().peekable();
+    let mut commas = commas.iter();
+    while let Some(item) = items.next() {
+        (item).write(stream, config);
+        if let Some(comma) = commas.next() {
+            // check if this comma can be dropped when trailing commas are disabled. Comma
+            // can be dropped iff this is the last item and the comma does not
+            // have a comment.
+            if config.trailing_commas || items.peek().is_some() || comma.has_comment() {
+                (comma).write(stream, config);
             }
-            _ if config.trailing_commas => {
-                stream.push_literal(",".to_string(), SyntaxKind::Comma);
-            }
-            _ => {}
+        } else if config.trailing_commas {
+            stream.push_literal(",".to_string(), SyntaxKind::Comma);
         }
         stream.end_line();
     }